# gen/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "contact": {
            "email": "support@codespace.dev",
            "name": "Codespace Operator Support",
            "url": "https://github.com/codespace-operator/codespace-operator"
        },
        "description": "REST API for managing Codespace sessions with RBAC and multi-provider authentication",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "termsOfService": "https://github.com/codespace-operator/codespace-operator/blob/main/TERMS.md",
        "title": "Codespace Operator API",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/v1/introspect/server": {
            "get": {
                "description": "Get server and cluster information",
                "parameters": [
                    {
                        "description": "Whether to discover namespaces (0 or 1)",
                        "in": "query",
                        "name": "discover",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ServerIntrospectionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "summary": "Server introspection",
                "tags": [
                    "user"
                ]
            }
        },
...
        "/auth/local/login": {
            "post": {
                "description": "Authenticate using username and password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/internal_server.LocalLoginRequest"
                            }
                        }
                    },
                    "description": "Login credentials",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "summary": "Local login",
                "tags": [
                    "authentication"
                ]
            }
        },
        "/auth/logout": {
            "get": {
                "description": "Clears session and, if OIDC is configured, redirects to provider end-session",
                "responses": {
                    "200": {
                        "description": "JSON {status: logged_out} if local",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "302": {
                        "description": "Redirect to provider end-session if available"
                    }
                },
                "summary": "Logout",
                "tags": [
                    "authentication"
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Refresh session token if valid",
                "responses": {
                    "204": {
                        "description": "No Content on success"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_server.ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "summary": "Refresh session",
                "tags": [
                    "authentication"
                ]
            }
        },
        "/auth/sso/login": {
            "get": {
                "description": "Redirects to OIDC provider for authentication",
                "parameters": [
                    {
                        "description": "Post-login relative redirect (safe, same-origin)",
                        "in": "query",
                        "name": "next",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to provider"
                    }
                },
                "summary": "Start OIDC login",
                "tags": [
                    "authentication"
                ]
            }
        },
...
    },
...
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer token authentication (JWT)",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            },
            "CookieAuth": {
                "description": "Session cookie authentication",
                "in": "cookie",
                "name": "codespace_session",
                "type": "apiKey"
            }
        },
        "schemas": {
            "common.AnchorMeta": {
                "properties": {
                    "name": {
                        "description": "release/app/deployment name (sanitized when used as label)",
                        "type": "string"
                    },
                    "namespace": {
                        "description": "namespace where the manager runs",
                        "type": "string"
                    },
                    "type": {
                        "description": "argo|helm|deployment|statefulset|daemonset|cronjob|job|pod|namespace|unresolved",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "internal_server.CasbinPermissions": {
                "properties": {
                    "namespaces": {
                        "$ref": "#/components/schemas/internal_server.NamespacePermissions"
                    }
                },
                "type": "object"
            },
            "internal_server.ClusterInfo": {
                "properties": {
                    "casbin": {
                        "$ref": "#/components/schemas/internal_server.CasbinPermissions"
                    },
                    "serverServiceAccount": {
                        "$ref": "#/components/schemas/internal_server.ServiceAccountInfo"
                    }
                },
                "type": "object"
            },
...
            "internal_server.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "Invalid request",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_server.LocalLoginRequest": {
                "description": "Local login credentials",
                "properties": {
                    "password": {
                        "example": "secretpassword",
                        "type": "string"
                    },
                    "username": {
                        "example": "alice",
                        "type": "string"
                    }
                },
                "required": [
                    "password",
                    "username"
                ],
                "type": "object"
            },
            "internal_server.LoginResponse": {
                "description": "Successful authentication response",
                "properties": {
                    "roles": {
                        "example": [
                            "editor",
                            "viewer"
                        ],
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "example": "alice",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "internal_server.NamespacePermissions": {
                "properties": {
                    "list": {
                        "type": "boolean"
                    },
                    "watch": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "internal_server.ServerIntrospectionResponse": {
                "properties": {
                    "capabilities": {
                        "$ref": "#/components/schemas/internal_server.SystemCapabilities"
                    },
                    "cluster": {
                        "$ref": "#/components/schemas/internal_server.ClusterInfo"
                    },
                    "instanceID": {
                        "type": "string"
                    },
                    "manager": {
                        "$ref": "#/components/schemas/common.AnchorMeta"
                    },
                    "namespaces": {
                        "$ref": "#/components/schemas/internal_server.ServerNamespaceInfo"
                    },
                    "version": {
                        "$ref": "#/components/schemas/internal_server.ServerVersionInfo"
                    }
                },
                "type": "object"
            },
            "internal_server.ServerNamespaceInfo": {
                "properties": {
                    "all": {
                        "description": "All namespaces (if discoverable)",
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "withSessions": {
                        "description": "Namespaces containing sessions",
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
            "internal_server.ServerVersionInfo": {
                "properties": {
                    "buildDate": {
                        "type": "string"
                    },
                    "gitCommit": {
                        "type": "string"
                    },
                    "version": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_server.ServiceAccountInfo": {
                "properties": {
                    "namespaces": {
                        "$ref": "#/components/schemas/internal_server.NamespacePermissions"
                    },
                    "session": {
                        "additionalProperties": {
                            "type": "boolean"
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "internal_server.SystemCapabilities": {
                "properties": {
                    "clusterScope": {
                        "description": "Displays cluster_scope: true or false",
                        "type": "boolean"
                    },
                    "multiTenant": {
                        "description": "Whether system supports multiple tenants",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    }
}